Add files via upload
Adventure Works Company

This dashboard helps the management to understand on the sales trends of the organization. It helps to have an overview on the Sales by by Fiscal year & month and also by region and country.  It also highlights the sales by category and the sales performance over the years. This will help the management to have an in depth understanding on the focus areas and the steps that needs to be taken to improve the sales in those areas.